backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
djangorestframework-simplejwt>=5.3
google-auth>=2.23
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1

